feat: add sector ranking, market env, notifications and enhance strategy/backtest engines

+++ backend/app/models/user.py
@@ -1,7 +1,7 @@
 """User and role models."""
 
 import datetime
-from sqlalchemy import Column, Integer, String, Boolean, DateTime, Text
+from sqlalchemy import Column, Integer, Float, String, Boolean, DateTime, Text
 from app.core.database import Base
 
 
@@ -15,7 +15,7 @@ class User(Base):
     display_name = Column(String(64), default="")
     role = Column(String(16), default="user")  # user, admin
     is_active = Column(Boolean, default=True)
-    initial_cash = Column(Integer, default=1000000)  # Default paper trading cash
+    initial_cash = Column(Float, default=1000000.0)  # Default paper trading cash
     created_at = Column(DateTime, default=datetime.datetime.utcnow)
     updated_at = Column(DateTime, default=datetime.datetime.utcnow, onupdate=datetime.datetime.utcnow)
     last_login = Column(DateTime, nullable=True)
